requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==3.2.6
django-allauth==0.45.0
django-discover-runner==1.0
django-rest-swagger==2.2.0
django-role-permissions==3.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
oauthlib==3.1.1
openapi-codec==1.3.2
Pillow==8.3.1
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
simplejson==3.17.4
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
